i386: kernel-loader: Add direct loading protected/long mode kernel APIs
Traditionally a PC follows the following booting steps:

QEMU -> BIOS -> bootloader -> kernel realmode code
    -> kernel protected/longmode code.

This process takes lots of time. For light virtualization platform we
let QEMU loads protected/long mode kernel directly, hence skipping
several steps and speeding up the whole booting. We do this by filling
the zero page per Linux booting protocol and then jumping to the kernel
protected/long mode entry. Registers and paging should also be put in
the correct state.
